requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5.4
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.14.60
botocore==1.17.60
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-admin-honeypot==1.1.0
django-allauth==0.42.0
django-autocomplete-light==3.5.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-fast-paginator==1.0.4
django-filter==2.2.0
django-friendship==1.9.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-redis==4.12.1
django-redis-cache==2.1.1
django-sendgrid-v5==0.9.0
django-storages==1.10
django-taggit==1.3.0
django-taggit-autosuggest==0.3.8
django-timezone-field==4.0
django-uuid-upload==0.0.2
djangorestframework==3.11.0
djutils==1.0.13
docutils==0.15.2
entrypoints==0.3
Faker==4.1.1
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
hashids==1.2.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
jmespath==0.10.0
joblib==0.16.0
jsonfield==3.1.0
kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.1
markdown2==2.3.9
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycryptodomex==3.9.8
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyquery==1.4.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2019.3
readtime==1.1.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sendgrid==6.4.6
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.0
starkbank-ecdsa==1.1.0
stream-python==3.5.2
swapper==1.1.2.post1
taggit-selectize==2.8.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.47.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.9
uuid==1.30
vine==1.3.0
whitenoise==5.2.0
wrapt==1.11.2
zipp==3.1.0
